# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Sales Academy Backend",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/simulator/sessions": {
      "post": {
        "tags": [
          "simulator"
        ],
        "summary": "Open Session",
        "operationId": "open_session_api_v1_simulator_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCreateResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/simulator/sessions/{session_id}/messages": {
      "post": {
        "tags": [
          "simulator"
        ],
        "summary": "Reply To Sales",
        "operationId": "reply_to_sales_api_v1_simulator_sessions__session_id__messages_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionMessageCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionMessageResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/simulator/sessions/{session_id}/finish": {
      "post": {
        "tags": [
          "simulator"
        ],
        "summary": "Close Session",
        "operationId": "close_session_api_v1_simulator_sessions__session_id__finish_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionFinishResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/audit-leads": {
      "post": {
        "tags": [
          "audit-leads"
        ],
        "summary": "Create Audit Lead",
        "operationId": "create_audit_lead_api_v1_audit_leads_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuditLeadCreateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditLeadDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "audit-leads"
        ],
        "summary": "List Audit Leads",
        "operationId": "list_audit_leads_api_v1_audit_leads_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AuditLeadStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditLeadListResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuditLeadCreateDto": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "clinic": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Clinic"
          },
          "contact": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Contact"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "source": {
            "type": "string",
            "maxLength": 64,
            "title": "Source",
            "default": "landing_audit_form"
          },
          "payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "name",
          "contact"
        ],
        "title": "AuditLeadCreateDto",
        "description": "Входящая заявка на аудит с лендинга."
      },
      "AuditLeadDto": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "clinic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clinic"
          },
          "contact": {
            "type": "string",
            "title": "Contact"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "status": {
            "$ref": "#/components/schemas/AuditLeadStatus"
          },
          "payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "contact",
          "source",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "AuditLeadDto"
      },
      "AuditLeadListResponseDto": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditLeadDto"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "AuditLeadListResponseDto"
      },
      "AuditLeadStatus": {
        "type": "string",
        "enum": [
          "new",
          "contacted",
          "qualified",
          "archived"
        ],
        "title": "AuditLeadStatus"
      },
      "CaseInfo": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "scenarioTitle": {
            "type": "string",
            "title": "Scenariotitle"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "scenarioTitle",
          "createdAt"
        ],
        "title": "CaseInfo"
      },
      "CompetencyAssessment": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "level": {
            "$ref": "#/components/schemas/ReportLevel"
          },
          "score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score"
          },
          "comment": {
            "type": "string",
            "title": "Comment"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceQuote"
            },
            "type": "array",
            "title": "Evidence"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "level",
          "score",
          "comment"
        ],
        "title": "CompetencyAssessment"
      },
      "CompetencyLevel": {
        "type": "string",
        "enum": [
          "Junior",
          "Middle",
          "Senior"
        ],
        "title": "CompetencyLevel"
      },
      "DevelopmentArea": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "comment": {
            "type": "string",
            "title": "Comment"
          },
          "actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Actions"
          }
        },
        "type": "object",
        "required": [
          "title",
          "comment"
        ],
        "title": "DevelopmentArea"
      },
...
      "DialogueSpeaker": {
        "type": "string",
        "enum": [
          "manager",
          "client",
          "assistant",
          "system"
        ],
        "title": "DialogueSpeaker"
      },
      "DialogueTurnAnalysis": {
        "properties": {
          "turnIndex": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Turnindex"
          },
          "speaker": {
            "$ref": "#/components/schemas/DialogueSpeaker"
          },
          "speakerLabel": {
            "type": "string",
            "title": "Speakerlabel"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "analysis": {
            "$ref": "#/components/schemas/TurnAnalysis"
          }
        },
        "type": "object",
        "required": [
          "turnIndex",
          "speaker",
          "speakerLabel",
          "text",
          "analysis"
        ],
        "title": "DialogueTurnAnalysis"
      },
      "EvaluationCompetencyRaw": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "level": {
            "$ref": "#/components/schemas/CompetencyLevel"
          },
          "argument": {
            "type": "string",
            "title": "Argument"
          },
          "quote": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Quote"
          },
          "recommendations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommendations"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "level",
          "argument",
          "quote",
          "recommendations"
        ],
        "title": "EvaluationCompetencyRaw"
      },
      "EvaluationResultRaw": {
        "properties": {
          "overall_level": {
            "$ref": "#/components/schemas/CompetencyLevel"
          },
          "overall_comment": {
            "type": "string",
            "title": "Overall Comment"
          },
          "overall_recommendations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Overall Recommendations"
          },
          "competencies": {
            "items": {
              "$ref": "#/components/schemas/EvaluationCompetencyRaw"
            },
            "type": "array",
            "title": "Competencies"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "overall_level",
          "overall_comment",
          "overall_recommendations",
          "competencies"
        ],
        "title": "EvaluationResultRaw"
      },
      "EvidenceQuote": {
        "properties": {
          "quote": {
            "type": "string",
            "title": "Quote"
          },
          "speaker": {
            "$ref": "#/components/schemas/DialogueSpeaker"
          },
          "turnIndex": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Turnindex"
          }
        },
        "type": "object",
        "required": [
          "quote",
          "speaker",
          "turnIndex"
        ],
        "title": "EvidenceQuote"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MessageRole": {
        "type": "string",
        "enum": [
          "learner",
          "customer"
        ],
        "title": "MessageRole"
      },
      "ParticipantInfo": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role",
            "default": "student"
          },
          "displayName": {
            "type": "string",
            "title": "Displayname",
            "default": "Ученик"
          }
        },
        "type": "object",
        "title": "ParticipantInfo"
      },
...
      "ReportLevel": {
        "type": "string",
        "enum": [
          "Trainee",
          "Junior",
          "Middle",
          "Senior"
        ],
        "title": "ReportLevel"
      },
...
      "ReportMeta": {
        "properties": {
          "generatedBy": {
            "type": "string",
            "title": "Generatedby",
            "default": "AI Sales Academy"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "dialogue_simulation"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "ru"
          },
          "fallback": {
            "type": "boolean",
            "title": "Fallback",
            "default": false
          }
        },
        "type": "object",
        "title": "ReportMeta"
      },
...
      "ReportSummary": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "headline": {
            "type": "string",
            "title": "Headline"
          },
          "overallLevel": {
            "$ref": "#/components/schemas/ReportLevel"
          },
          "overallScore": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Overallscore"
          },
          "shortResume": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Shortresume"
          }
        },
        "type": "object",
        "required": [
          "title",
          "headline",
          "overallLevel",
          "overallScore"
        ],
        "title": "ReportSummary"
      },
...
      "SalesDialogueReportV2": {
        "properties": {
          "reportVersion": {
            "type": "string",
            "title": "Reportversion",
            "default": "2.0"
          },
          "case": {
            "$ref": "#/components/schemas/CaseInfo"
          },
          "participant": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ParticipantInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "summary": {
            "$ref": "#/components/schemas/ReportSummary"
          },
          "competencies": {
            "items": {
              "$ref": "#/components/schemas/CompetencyAssessment"
            },
            "type": "array",
            "title": "Competencies"
          },
          "dialogueAnalysis": {
            "items": {
              "$ref": "#/components/schemas/DialogueTurnAnalysis"
            },
            "type": "array",
            "title": "Dialogueanalysis"
          },
          "strengths": {
            "items": {
              "$ref": "#/components/schemas/Strength"
            },
            "type": "array",
            "title": "Strengths"
          },
          "developmentAreas": {
            "items": {
              "$ref": "#/components/schemas/DevelopmentArea"
            },
            "type": "array",
            "title": "Developmentareas"
          },
          "nextSteps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Nextsteps"
          },
          "meta": {
            "$ref": "#/components/schemas/ReportMeta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "case",
          "summary"
        ],
        "title": "SalesDialogueReportV2"
      },
...
      "SessionCreateDto": {
        "properties": {
          "scenario_id": {
            "type": "string",
            "title": "Scenario Id"
          },
          "difficulty": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficulty"
          }
        },
        "type": "object",
        "required": [
          "scenario_id"
        ],
        "title": "SessionCreateDto"
      },
      "SessionCreateResponseDto": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "message": {
            "$ref": "#/components/schemas/SessionMessageDto"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "status",
          "message"
        ],
        "title": "SessionCreateResponseDto"
      },
      "SessionFinishResponseDto": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "evaluation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvaluationResultRaw"
              },
              {
                "type": "null"
              }
            ]
          },
          "report_v2": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SalesDialogueReportV2"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "status"
        ],
        "title": "SessionFinishResponseDto"
      },
      "SessionMessageCreateDto": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "SessionMessageCreateDto"
      },
      "SessionMessageDto": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "role": {
            "$ref": "#/components/schemas/MessageRole"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "role",
          "text"
        ],
        "title": "SessionMessageDto"
      },
      "SessionMessageResponseDto": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "rude": {
            "type": "string",
            "title": "Rude"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/SessionMessageDto"
            },
            "type": "array",
            "title": "Messages"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "status",
          "rude",
          "confidence",
          "messages"
        ],
        "title": "SessionMessageResponseDto"
      },
      "SessionStatus": {
        "type": "string",
        "enum": [
          "active",
          "finished"
        ],
        "title": "SessionStatus"
      },
      "Strength": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "comment": {
            "type": "string",
            "title": "Comment"
          },
          "evidence": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence"
          }
        },
        "type": "object",
        "required": [
          "title",
          "comment"
        ],
        "title": "Strength"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}